backend/pyproject.toml
[project]
name = "energiemanager"
version = "0.1.0"
description = "Intelligentes Energiemanagementsystem"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "redis>=5.2",
    "celery>=5.4",
    "httpx>=0.28",
    "websockets>=14.0",
    "numpy>=2.0",
    "pandas>=2.2",
    "scikit-learn>=1.6",
    "xgboost>=2.1",
    "pulp>=2.9",
    "pymodbus>=3.7",
    "paho-mqtt>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "httpx>=0.28",
    "ruff>=0.8",
]
ml = [
    "torch>=2.5",
    "optuna>=4.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
